+++ src/app/product/[productId]/actions.ts
@@ -3,12 +3,12 @@
 import { cookies } from 'next/headers';
 import { ProductDetailType } from '@/types/types.ts';
 
-const fetchProductData = async (
+const cookieStore = cookies();
+const accessToken = cookieStore.get('accessToken');
+
+export const fetchProductData = async (
   productId: string,
 ): Promise<ProductDetailType> => {
-  const cookieStore = cookies();
-  const accessToken = cookieStore.get('accessToken');
-
   const headers: HeadersInit = {
     'Content-Type': 'application/json',
   };
@@ -31,4 +31,27 @@ const fetchProductData = async (
   return response.json();
 };
 
-export default fetchProductData;
+export async function fetchUserInfo() {
+  if (!accessToken) {
+    console.error('Access token is missing');
+    return null;
+  }
+
+  try {
+    const response = await fetch(
+      `${process.env.NEXT_PUBLIC_API_URL_HOST}/users/me`,
+      {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${accessToken.value}`,
+        },
+      },
+    );
+
+    return await response.json();
+  } catch (error) {
+    console.error('Error during fetch operation:', error);
+    return null;
+  }
+}
