+++ src/components/ProductionReview/index.tsx
@@ -1,31 +1,23 @@
 import { useQuery } from '@tanstack/react-query';
 import cx from '@/components/ProductionReview/cx.ts';
-import {
-  getReview,
-  fetchUserInfo,
-} from '@/components/ProductionReview/actions.ts';
+import { getReview } from '@/components/ProductionReview/actions.ts';
 import ReviewItem from '@/components/ReviewItem';
-import { Review } from '@/types/types.ts';
+import { Review, UserDetail } from '@/types/types.ts';
 
 interface Props {
   productId: string;
   order: string;
+  me: UserDetail;
 }
 export default function ProductionReview({
   productId,
   order,
+  me,
 }: Readonly<Props>) {
   const { data: reviewData } = useQuery({
     queryKey: ['reviews', productId, order],
     queryFn: () => getReview(productId, order),
   });
-
-  const { data: me } = useQuery({
-    queryKey: ['me'],
-    queryFn: () => fetchUserInfo(),
-  });
-
-  console.log(reviewData);
 
   return (
     <div className={cx('container')}>
